+++ firmware/console/binary/tunerstudio.mk
@@ -1,6 +1,7 @@
 
 TUNERSTUDIO_SRC_CPP = $(PROJECT_DIR)/console/binary/tunerstudio_io.cpp \
 	$(PROJECT_DIR)/console/binary/tunerstudio_io_serial.cpp \
+	$(PROJECT_DIR)/console/binary/tunerstudio_io_serial_ports.cpp \
 	$(PROJECT_DIR)/console/binary/ts_can_channel.cpp \
 	$(PROJECT_DIR)/console/binary/tunerstudio.cpp \
 	$(PROJECT_DIR)/console/binary/tunerstudio_commands.cpp \
